loading controls using reflection, fixed reflection initialization approach

+++ engine/UnitTest.Core/src/core/collection/ListTest.cpp
@@ -432,13 +432,13 @@ public:
         Timer timer;
         timer.tick();
         List<uint32_t> list;
-        for (size_t i = 0; i < 100000; i++)
+        for (uint32_t i = 0; i < 100000; i++)
             list.add(i);
         timer.tick();
         auto timeList = timer.FrameTimeMs;
         timer.tick();
         std::vector<uint32_t> vector;
-        for (size_t i = 0; i < 100000; i++)
+        for (uint32_t i = 0; i < 100000; i++)
             vector.push_back(i);
         timer.tick();
         auto timeVector = timer.FrameTimeMs;
